`InternalDeployAccount` constructor should receive chain id as Felt (#546)
The `StarknetChainId` type is not generic enough to represent chain IDs
for custom chains (and even integration is missing). This change converts
`InternalDeployAccount::new()` to receive the chain ID as a felt (like
other transaction types already do).

It also removes the `chain_id` member of the struct and simply re-uses
the transaction hash calculated in the constructor in
`get_execution_context()`.

Co-authored-by: Santiago Pittella <87827390+SantiagoPittella@users.noreply.github.com>